fix: 修正代码

+++ src/main/java/com/jirengu/TwoSum.java
@@ -3,10 +3,10 @@ package com.jirengu;
 public class TwoSum {
     public static void main(String[] args) {
         //请修复下面的代码使得能正确计算打印Sum: 214
-        byte num1 = 104;
-        byte num2 = 110;
-        byte sum;
-        sum = (byte)(num1 + num2);
+        short num1 = 104;
+        short num2 = 110;
+        short sum;
+        sum = (short)(num1 + num2);
         System.out.print("Sum: " + sum);
     }
 }
